Added hacky access control

+++ requirementsTracker/rct/views.py
@@ -10,6 +10,14 @@ from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
 
 
+def is_member(project, user_id):
+    for m in project.members.all():
+        if m.id == user_id:
+            return True
+
+    return False
+
+
 def user_login(request):
 	context = RequestContext(request)
 	if request.method == 'POST':
@@ -36,7 +44,12 @@ def user_login(request):
 def index(request):
     context = RequestContext(request)
 
-    projects = Project.objects.all()
+    projects = []
+    for p in Project.objects.all():
+        if is_member(p, request.user.id):
+            projects.append(p)
+            break
+
     context_dict = {'projects' : projects}
 
     for p in projects:
@@ -47,8 +60,12 @@ def index(request):
 def edit_project(request, url):
     context = RequestContext(request)
 
+    project = Project.objects.get(url=url)
+    if not is_member(project, request.user.id):
+        HttpResponse("Access denied")
+
     if request.method == 'POST':
-        form = EditProjectForm(request.POST, instance=Project.objects.get(url=url))
+        form = EditProjectForm(request.POST, instance=project)
         if form.is_valid():
                 project = form.save()
 
@@ -80,43 +97,50 @@ def view_project(request, url):
     context_dict = {}
 
     try:
-        project = Project.objects.get(url__iexact=url)
+        project = Project.objects.get(url=url)
+        if not is_member(project, request.user.id):
+            return HttpResponse("Access denied")
+
         context_dict['project'] = project
         context_dict['form']  = EditProjectForm(instance=project)
     except Project.DoesNotExist:
         return HttpResponseNotFound('<h1>Project not found</h1>')
-    
+
     try:
         for key in ('must', 'should', 'could', 'would'):
             context_dict[key] = context_dict['project'].task_set.filter(category=key[0].upper()).all()
     except Exception:
         # guess there are no tasks
         pass    
-    
+
     return render_to_response('rct/projects/view.html', context_dict, context)
 
 @login_required
 def add_task(request, url):
 
-	context = RequestContext(request)
-	context_dict = {}
-	try:
-		context_dict['project'] = Project.objects.get(url__iexact=url)
-	except Project.DoesNotExist:
-		pass
-	
-	if request.method == 'POST':
-		form = TaskForm(request.POST)
-		if form.is_valid():
-			task = form.save(commit=False)
-			task.project = context_dict['project']
-			task.save()
+    context = RequestContext(request)
+    context_dict = {}
+    try:
+        project = Project.objects.get(url=url)
+        if not is_member(project, request.user.id):
+            return HttpResponse("Access denied")
 
-			return HttpResponseRedirect(reverse('rct.views.index') + "project/" + url)
-	else:
-		form = TaskForm()
-	context_dict['form'] = form
-	return render_to_response('rct/tasks/create.html', context_dict, context)
+        context_dict['project'] = project
+    except Project.DoesNotExist:
+        pass
+
+    if request.method == 'POST':
+        form = TaskForm(request.POST)
+        if form.is_valid():
+            task = form.save(commit=False)
+            task.project = context_dict['project']
+            task.save()
+
+            return HttpResponseRedirect(reverse('rct.views.index') + "project/" + url)
+    else:
+        form = TaskForm()
+    context_dict['form'] = form
+    return render_to_response('rct/tasks/create.html', context_dict, context)
 
 @login_required
 def projectBoard(request):
@@ -156,5 +180,10 @@ def delete_task(id):
 
 @login_required
 def delete_project(request,url):
+
+    project = Project.objects.get(url=url)
+    if not is_member(project, request.user.id):
+        return HttpResponse("Access denied")
+
     Project.objects.filter(url=url).delete()
     return HttpResponseRedirect('/rct/')
